feat: Move delete action code to table's row
