+++ static/app/views/settings/project/tempest/EmptyState.tsx
@@ -24,18 +24,18 @@ import type {TempestCredentials} from 'sentry/views/settings/project/tempest/typ
 
 interface EmptyStateProps {
   hasWriteAccess: boolean;
-  isRemoving: boolean;
-  onRemoveCredential: (data: {id: number}) => void;
+  isLoading: boolean;
+  onRemoveCredentialSuccess: () => void;
   project: Project;
   tempestCredentials?: TempestCredentials[];
 }
 
 export default function EmptyState({
   project,
+  isLoading,
   tempestCredentials,
-  isRemoving,
   hasWriteAccess,
-  onRemoveCredential,
+  onRemoveCredentialSuccess,
 }: EmptyStateProps) {
   const navigate = useNavigate();
   const location = useLocation();
@@ -93,6 +93,7 @@ export default function EmptyState({
                     t('Created By'),
                     '',
                   ]}
+                  isLoading={isLoading}
                   isEmpty={!tempestCredentials?.length}
                   emptyMessage={t('No credentials found')}
                   emptyAction={<AddCredentialsButton project={project} />}
@@ -101,8 +102,9 @@ export default function EmptyState({
                     <CredentialRow
                       key={credential.id}
                       credential={credential}
-                      isRemoving={isRemoving}
-                      removeCredential={hasWriteAccess ? onRemoveCredential : undefined}
+                      hasWriteAccess={hasWriteAccess}
+                      onRemoveCredentialSuccess={onRemoveCredentialSuccess}
+                      project={project}
                     />
                   ))}
                 </StyledPanelTable>
